requirements.txt
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
